LayoutTests/http/tests/security/xssAuditor/property-escape-comment-03-expected.txt

CONSOLE MESSAGE: line 3: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-property.pl?clutter=%3cdiv%3e&q=%22%20%22%20onload=alert(3)%3C!--' because its source code was found within the request. The auditor was enabled because the server did not send an 'X-XSS-Protection' header.
